- Fixing the bug passing a dictionary (2019_04_24):
Just so that you know the changes, i was passing a dictionary (from the output of LOCAL and PEAK FDR) which was huge (in the case of PESA i suppose, more than 6 GB), and that is why i was crashing when i was passing the dict to next method.
Now program does not save it anymore, i write a big file out of it and than split it in every experiment and than read it when required in PSM to peak assignment. you will see a new method in localFDR_median.py called def ProduceFileforEveryExperiment.

this method first produces a "Peak_and_Slope_FDRfile.txt" and than later splits into "Peak_and_Slope_FDRfile.txt-Smallchunk.txt".