From: Elena Bonzón Kulichenko Sent: Monday, April 29, 2019 9:49 AM
Dear Navratan,
Thanks again for the prompt answer! I see what you mean.
However, if we are running the PTMcorrector in the 3-d modality, that is, we want to do only Isotopic correction, it should be done only with the PSMs included in the peaks from the peak list, which we have generated vary carefully. The orphans in this case would be considered as noise and should not be taken into account.