Add new scripts: isotopicCorrection
The algorithm works in this way. 1) it takes the similar sequence together 2) sorts them according to DM (smallest to largest). 3) starts from smallest and checks for C13 and sulphur in upto two jumps.

in the case you mentioned: it first encounters  -0.985242 which s not a c13 or a sulphar from its previous DM which is -1.003864 ... and so on.

to answer Jesus's question: i think it better to do it after FDR correction considering the change you proposed yesterday. before FDR you might have peaks (may be false one), which is bigger than the consecutive  peak, and it would perform a correction, but in realty after FDR the abundance of big peak might change (considering it might be a false one).

i think it will better i talk for a minute over skype and clear it out, --> writing it in mail mail might be bit confusing .
call me on Skype if you are free ?